[OCTREE] Replace unsigned by size_t

+++ src/octree.cc
@@ -3,7 +3,7 @@
 # include <cassert>
 # include <cmath>
 
-allrgb::Octree::Octree(unsigned depth)
+allrgb::Octree::Octree(size_t depth)
   : data_((std::pow(8ul, depth) - 1) / (8ul - 1))
   , nb_leaves_(std::pow(8ul, depth - 1))
 {
@@ -17,13 +17,13 @@ allrgb::Octree::init_()
 }
 
 void
-allrgb::Octree::init_(unsigned long index, unsigned long sub_leaves)
+allrgb::Octree::init_(size_t index, size_t sub_leaves)
 {
   data_.at(index) = sub_leaves;
 
   if (sub_leaves <= 1)
     return;
 
-  for (unsigned i = 0; i < 8; ++i)
+  for (size_t i = 0; i < 8; ++i)
     init_(8 * index + 1 + i, sub_leaves / 8);
 }
